Resolved issues

+++ server/src/modules/community-event/service.ts
@@ -6,17 +6,20 @@ import { updateCommunityEventReq } from './validations';
 import communityEventSerializer from './serializer';
 
 const communityEventService = {
-  findAll: async () => {
+  findAll: async (userId: string) => {
     return await prisma.communityEvent.findMany({
+      where: {
+        organizerId: userId,
+      },
       include: {
         eventType: true,
         organizer: true,
       },
     });
   },
-  findById: async (communityEventId: string) => {
-    const exisitngEvent = await prisma.communityEvent.findUnique({
-      where: { id: communityEventId },
+  findById: async (communityEventId: string, userId: string) => {
+    const exisitngEvent = await prisma.communityEvent.findFirst({
+      where: { id: communityEventId, organizerId: userId },
       include: {
         eventType: true,
         organizer: true,
@@ -28,7 +31,7 @@ const communityEventService = {
         `Community Event with id ${communityEventId} not found`,
       );
     }
-    console.log(exisitngEvent)
+    console.log(exisitngEvent);
     return exisitngEvent;
   },
   create: async (params: Prisma.CommunityEventCreateInput, userId: string) => {
@@ -58,9 +61,10 @@ const communityEventService = {
   updateById: async (
     communityEventId: string,
     updateData: z.infer<typeof updateCommunityEventReq>['body'],
+    userId: string,
   ) => {
-    const existingCommunityEvent = !!(await prisma.communityEvent.findUnique({
-      where: { id: communityEventId },
+    const existingCommunityEvent = !!(await prisma.communityEvent.findFirst({
+      where: { id: communityEventId, organizerId: userId },
     }));
 
     if (!existingCommunityEvent) {
@@ -84,9 +88,9 @@ const communityEventService = {
    * @param communityEventId
    * @returns
    */
-  deleteById: async (communityEventId: string) => {
-    const existingCommunityEvent = !!(await prisma.communityEvent.findUnique({
-      where: { id: communityEventId },
+  deleteById: async (communityEventId: string, userId: string) => {
+    const existingCommunityEvent = !!(await prisma.communityEvent.findFirst({
+      where: { id: communityEventId, organizerId: userId },
     }));
 
     if (!existingCommunityEvent) {
